Fixes clippy warnings and dead_code warnings

+++ server/src/http_server/hypermedia/routes/ups.rs
@@ -173,7 +173,7 @@ impl<'a> UpsInfoTemplate<'a> {
       ..
     } = template
     {
-      template.power = Some((pw * f64::from(ld)) / 100.0_f64);
+      template.power = Some((pw * ld) / 100.0_f64);
     };
 
     template
